Fixed: apply [3660] to mangos.sql, list sql update in makefile.am
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3661 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/sql/mangos.sql
@@ -376,7 +376,7 @@ UNLOCK TABLES;
 
 DROP TABLE IF EXISTS `character_kill`;
 CREATE TABLE `character_kill` (
-  `guid` smallint(4) unsigned NOT NULL default '0' COMMENT 'Global Unique Identifier',
+  `guid` int(11) unsigned NOT NULL default '0' COMMENT 'Global Unique Identifier',
   `creature_template` int(11) unsigned NOT NULL default '0' COMMENT 'Creature Identifier',
   `honor` float NOT NULL default '0',
   `date` int(11) unsigned NOT NULL default '0',
